prepared for npm launch
