+++ package.json
@@ -1,0 +1,27 @@
+{
+  "name": "cadet-parser",
+  "version": "0.1.0",
+  "description": "a mvp markdown parser for use with Cadet",
+  "main": "index.js",
+  "directories": {
+    "example": "example"
+  },
+  "scripts": {
+    "test": "./node_modules/mocha/bin/mocha test",
+    "start": "./index.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/timcmiller/simple_md_parser.git"
+  },
+  "author": "",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/timcmiller/simple_md_parser/issues"
+  },
+  "homepage": "https://github.com/timcmiller/simple_md_parser#readme",
+  "devDependencies": {
+    "chai": "^3.4.1",
+    "mocha": "^2.3.4"
+  }
+}
